Refactor: 참여중인 로드맵 코드 리팩토링
