+++ src/components/Roadmap/MyRoadmap/MyRoadmapList/style.ts
@@ -1,0 +1,64 @@
+import SliderList from 'react-slick';
+import styled from '@emotion/styled';
+import SkeletonBox from '@/components/common/Skeleton';
+
+export const Slider = styled(SliderList)`
+  position: relative;
+
+  @media (max-width: 550px) {
+    width: 250px;
+    margin: 0 auto;
+  }
+
+  & > div {
+    overflow: hidden;
+  }
+
+  & > div > div {
+    display: flex;
+    justify-content: space-evenly;
+    gap: 10px;
+  }
+
+  & > button:nth-of-type(1) {
+    position: absolute;
+    top: 100px;
+    left: -35px;
+    width: 20px;
+    height: 40px;
+    background-image: url('/assets/icons/ic_chevronLeftBlack.svg');
+    background-repeat: no-repeat;
+    background-size: contain;
+    font-size: 0;
+  }
+
+  & > button:nth-of-type(2) {
+    position: absolute;
+    bottom: 90px;
+    right: -30px;
+    width: 20px;
+    height: 40px;
+    background-image: url('/assets/icons/ic_chevronRightBlack.svg');
+    background-repeat: no-repeat;
+    background-size: contain;
+    font-size: 0;
+  }
+`;
+
+export const SkeletonRoot = styled.section`
+  display: flex;
+  gap: 10px;
+`;
+export const Skeleton = styled(SkeletonBox)`
+  width: 211px;
+  height: 230px;
+  margin: 0 auto;
+`;
+
+export const EmptyRoot = styled.section`
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 30px;
+  margin: 55px auto 55px;
+`;
