+++ src/components/Roadmap/RoadmapCodeModal/style.ts
@@ -1,0 +1,13 @@
+import styled from '@emotion/styled';
+
+export const RoadmapCodeModal = styled.article`
+  display: flex;
+  flex-direction: column;
+  gap: 0.5rem;
+`;
+
+export const RoadmapCodeForm = styled.form`
+  display: flex;
+  flex-direction: column;
+  gap: 0.5rem;
+`;
